Fast Pancake Object Handlers
Custom class and object handlers for all Pancake classes. The new
handlers allow faster property and method accesses. Changes include that
every property and method is forced to be public, no matter how it was
declared. __get(), __set(), dynamic properties and such things are not
supported. isset() on a property will return true also when it is NULL.
In case you would like to use the Pancake object handlers for your own
classes in a userscript, you may call Pancake\makeFastClass(string
$className). All objects of the class that are created afterwards will
use the Pancake object handlers instead of the PHP/Zend ones.